feat:CTTIIDPS-8290 add metrics percentage consumptions for cosmos db

+++ cloud_development/app/domain/CosmosDbMetric.py
@@ -20,7 +20,7 @@ class CosmosDbMetric:
     autoscaleMaxThroughputProposed:Decimal
     azureCostsProposed:Decimal
     spikesThroughputProposed:int
-    percentageUtilizationProposed:int
+    percentageUtilizationProposed:Decimal
 
     averageHalfRequestUnits:Decimal
     maximumRequestUnits:Decimal
